Rename test-package functional -> feature
