+++ .gitignore
@@ -8,3 +8,4 @@ dependency-reduced-pom.xml
 .terraform
 terraform.tfstate*
 out/
+.claude/*
