+++ db-scheduler/src/test/java/com/github/kagkarlsson/scheduler/feature/DeleteUnresolvedTest.java
@@ -1,4 +1,4 @@
-package com.github.kagkarlsson.scheduler.functional;
+package com.github.kagkarlsson.scheduler.feature;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
